WIP: use localhost6 for all service URLs, move all installer downloads to their own provisioner
This CentOS VM seems to prefer IPv6, so let's use localhost6 as the
domain for all services instead of localhost (which is IPv4).

The installer downloads are pretty fragile, moving them all to their own
provisioner lets us focus on gathering them in the beginning, and keeps
the installers around in case we need them again.